[e2e] Fix token warnings (gnosis/cowswap#114)
Fix token warnings. Part of #104 

Now working!
![image](https://user-images.githubusercontent.com/2352112/105985707-219c7b00-609c-11eb-8150-a1864fdc9fb2.png)

There was several issues. First, the test run in Rinkeby, and the token was changed to one that doesn't exist in that network. This PR reverts that change.

Additionally, for convenient we created `REACT_APP_DISABLE_TOKEN_WARNING` env var, which save us a click every time we load locally the tokens by URL. We do this constantly in local since it's so slow

Now, the default is to not set this env var. Although `yarn start` will add it for convenience.

I added yarn start:default, in case we want to start the app with the default (no env var override). Useful for example if we serve Cypress directly (insted of using `yarn serve`)