Paraswap price (gnosis/cowswap#791)
This PR generalize the price quote querying, so instead of only using one souce (our API), now we can use more. 

It includes Paraswap price.

Part of #780 

## About networks
Paraswap only has Mainnet in common with us. 

Their library need to be instanciated for a specific network. This network later will be used to pass to their API endpoint.
For this reason, the app keeps a cache of one library per network.

## Additional changes
The operator was adapting the token addresses to what the API expects. Basically, it has to translate native ETH into WETH.
The same logic is required for Paraswap. So this logic was simplified and moved to one util.

## Performance
Paraswap solver is slower, this adds a delay to the query, although feels acceptable for now, since we already have the loading indicator and is not that bad

## Not included
- Timeout, this will be addressed in other PRs

## Testing
Some keys for testing. This PR modifies getting the prices, so this part should be tested more heavily.
Special attention to:
- Sell Eth trades in all networks
- sell/buy in all networks test are required
- Try to buy ETH in all networks
- Compare price to CowSwap / Paraswap
- Review params of the query to paraswap: buy/sell orders

Expect that ONLY in mainnet it will query to Paraswap. We should get better or equal prices than CowSwap, specially for bigger amounts.