Update wrong word in COW FAQ (gnosis/cowswap#908)
In the line 253, for the question: To what does the term Coincidence of Wants  (CoWs) refer? We have "This leads to better prices for the individual traders (because traditionally market makers add a fee — referred to as spread — for their surface)". This PR changes the word surface for the correct word, "service"