[Affiliate] Copy link component (gnosis/cowswap#968)
# Summary
Closes #851 

Edit Copy component in order to allow copying a link either by clicking over the icon or address. 

As part of the discussion [here](https://github.com/gnosis/cowswap/pull/952#pullrequestreview-708958950) I decided to split this issue in 2 PRs (this is the one related to the implementation and then there will be another related to the integrations once the affiliate page is merged)


# Testing steps
Since it is not a integration and given the fact we don't have Storybook to list components, there is no visual feedback until it is integrated into the affiliate page.